Moved and added DaceFileDiscovery to startup
