+++ Ready4Balfolk.Domain/Models/Tracks/DanceFileEntry.cs
@@ -1,0 +1,3 @@
+namespace Ready4Balfolk.Domain.Models.Tracks;
+
+public record DanceFileEntry(string FileName, string Dance = "Unknown");
