+++ Ready4Balfolk.Domain/Models/Tracks/DiscoveryPattern.cs
@@ -1,0 +1,9 @@
+namespace Ready4Balfolk.Domain.Models.Tracks;
+
+public record DiscoveryPattern(ICollection<string> Pattern)
+{
+    public static readonly DiscoveryPattern DefaultDefault = new(["%d - %a - %t.%x"]);
+    public static readonly DiscoveryPattern ExtendedDefault = new(["%a", "%l", "%n - %t.%x"]);
+
+    public static implicit operator string[](DiscoveryPattern pattern) => [.. pattern.Pattern];
+}
