+++ Ready4Balfolk.Domain/Models/Tracks/PatternSegment.cs
@@ -1,0 +1,13 @@
+namespace Ready4Balfolk.Domain.Models.Tracks;
+
+public enum PatternSegment
+{
+    Artist,
+    Album,
+    Title,
+    TrackNumber,
+    Year,
+    Genre,
+    Dance,
+    Extension
+}
