+++ Ready4Balfolk.Domain/Services/Tracks/Discovery/IDanceFileDiscoveryService.cs
@@ -1,0 +1,8 @@
+using System.IO.Abstractions;
+
+namespace Ready4Balfolk.Domain.Services.Tracks.Discovery;
+
+public interface IDanceFileDiscoveryService
+{
+    Dictionary<string, string> Matches(IDirectoryInfo directoryInfo);
+}
