+++ Ready4Balfolk.Tests/Unit/DanceFileDiscoveryServiceTests.cs
@@ -1,9 +1,11 @@
 using System.IO.Abstractions.TestingHelpers;
+using Ready4Balfolk.Domain.Models.Tracks;
 using Ready4Balfolk.Domain.Services.Tracks;
+using Ready4Balfolk.Domain.Services.Tracks.Discovery;
 
 namespace Ready4Balfolk.Tests.Unit;
 
-public class DanceFileDiscoveryTests
+public class DanceFileDiscoveryServiceTests
 {
     [Fact]
     public void BasicWriteReadTest()
@@ -13,7 +15,7 @@ public class DanceFileDiscoveryTests
         var directoryInfo = mockFileSystem.DirectoryInfo.New("TestDirectory");
         directoryInfo.Create();
 
-        var dfd = new DanceFileDiscovery(mockFileSystem);
+        var dfd = new DanceFileDiscoveryService(new DanceFileService(mockFileSystem));
 
         ICollection<DanceFileEntry> content = [new DanceFileEntry("Test 1", "Dance 1")];
         dfd.Write(directoryInfo, content);
@@ -34,7 +36,7 @@ public class DanceFileDiscoveryTests
         var directoryInfo = mockFileSystem.DirectoryInfo.New("TestDirectory");
         directoryInfo.Create();
 
-        var dfd = new DanceFileDiscovery(mockFileSystem);
+        var dfd = new DanceFileDiscoveryService(new DanceFileService(mockFileSystem));
 
         var readContent = dfd.Matches(directoryInfo);
         Assert.Empty(readContent);
@@ -46,7 +48,7 @@ public class DanceFileDiscoveryTests
         var mockFileSystem = new MockFileSystem();
 
         var dir = mockFileSystem.DirectoryInfo.New("TestDirectory");
-        var dfd = new DanceFileDiscovery(mockFileSystem);
+        var dfd = new DanceFileDiscoveryService(new DanceFileService(mockFileSystem));
 
         Assert.Throws<DirectoryNotFoundException>(() => dfd.Matches(dir));
     }
