+++ Ready4Balfolk.Tests/Unit/PatternMatcherTests.cs
@@ -1,5 +1,7 @@
 using System.IO.Abstractions.TestingHelpers;
+using Ready4Balfolk.Domain.Models.Tracks;
 using Ready4Balfolk.Domain.Services.Tracks;
+using Ready4Balfolk.Domain.Services.Tracks.Discovery.Matching;
 
 namespace Ready4Balfolk.Tests.Unit;
